requirements.txt
# =============================================================================
# PSC_Agents 项目依赖配置
# Perovskite Solar Cell Multi-Agent Research System
# =============================================================================

# -----------------------------------------------------------------------------
# 核心依赖 - Core Dependencies
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0          # 环境变量管理

# -----------------------------------------------------------------------------
# MCP (Model Context Protocol) - Agent 工具调用协议
# -----------------------------------------------------------------------------
mcp>=1.0.0                    # MCP SDK
httpx>=0.27.0                 # HTTP 异步客户端 (MCP SSE 连接)

# -----------------------------------------------------------------------------
# LangChain 生态 - LLM 框架
# -----------------------------------------------------------------------------
langchain-core>=0.3.0         # LangChain 核心
langchain-openai>=0.2.0       # OpenAI / 代理 API 支持
langchain-anthropic>=0.2.0    # Anthropic Claude 支持
langchain-google-genai>=2.0.0 # Google Gemini 支持 (可选)


# -----------------------------------------------------------------------------
# LangGraph - 工作流编排
# -----------------------------------------------------------------------------
langgraph>=0.2.0              # 多智能体工作流图

# -----------------------------------------------------------------------------
# 数据处理 - Data Processing
# -----------------------------------------------------------------------------
pandas>=2.0.0                 # 数据框处理
numpy>=1.24.0                 # 数值计算

# -----------------------------------------------------------------------------
# 机器学习 - Machine Learning
# -----------------------------------------------------------------------------
scikit-learn>=1.3.0           # ML 算法、数据预处理
joblib>=1.3.0                 # 模型持久化

# -----------------------------------------------------------------------------
# 材料学与化学计算 - Materials Science & Chemistry
# -----------------------------------------------------------------------------
pymatgen>=2024.1.1            # 材料学计算：化学计量分析、电荷平衡、结构处理
ase>=3.22.0                   # 原子模拟环境：晶体结构读写、CIF 文件处理
rdkit>=2023.9.1               # 化学信息学：有机分子分析、疏水性计算 (pip install rdkit)

# -----------------------------------------------------------------------------
# 可解释AI - Explainable AI
# -----------------------------------------------------------------------------
shap>=0.44.0                  # SHAP 特征重要性分析

# -----------------------------------------------------------------------------
# 可视化 - Visualization
# -----------------------------------------------------------------------------
matplotlib>=3.7.0             # 静态图表
plotly>=5.18.0                # 交互式3D晶体结构可视化
seaborn>=0.13.0               # 统计可视化 (可选，用于实验绘图)